:pencil: Removed unneeded double keyword

+++ code/force_models/src/RudderForceModel.cpp
@@ -220,7 +220,7 @@ ssc::kinematics::Vector6d RudderForceModel::get_rudder_force(const BodyStates& s
 Wrench RudderForceModel::get_force(const BodyStates& states, const double t, const EnvironmentAndFrames& env, const std::map<std::string,double>& commands) const
 {
     const ssc::kinematics::Vector6d propeller_force = propulsion.get_force(states, t, env, commands).to_vector();
-    const ssc::kinematics::Vector6d rudder_force = get_rudder_force(states, t, env, commands, (double)propeller_force.norm());
+    const ssc::kinematics::Vector6d rudder_force = get_rudder_force(states, t, env, commands, propeller_force.norm());
     const std::string frame = translation_from_rudder_to_propeller.get_frame();
     const ssc::kinematics::Wrench prop(frame, propeller_force);
     const ssc::kinematics::Wrench rudder_wrench(ssc::kinematics::Point(frame,0,0,0), rudder_force);
